requirements.txt
# =====================================================
# BOT E-CAC - DEPENDÊNCIAS PYTHON
# =====================================================

# FastAPI e servidor
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Base de dados
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
alembic==1.14.0

# HTTP e APIs
httpx==0.28.0
requests==2.32.3
aiohttp==3.11.10

# Autenticação e segurança
cryptography==43.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Certificado digital e assinatura
pyOpenSSL==24.3.0

# Validação de dados
pydantic==2.10.2
pydantic[email]==2.10.2
python-cpf==1.0.0
validate-docbr==1.10.0

# Configuração e ambiente
python-dotenv==1.0.1
pydantic-settings==2.6.1

# Logs e monitoramento
loguru==0.7.2
prometheus-client==0.21.0

# Utilidades
python-dateutil==2.9.0
pytz==2024.2
uuid7==0.1.0

# Desenvolvimento
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.0

# CORS
python-cors==1.7.0 